requirements.txt
Django==4.1.3
djangorestframework==3.14.0
Pillow==9.4.0
PyJWT==2.6.0
